lesson4/seminar/step4_monitoring/pyproject.toml
[tool.poetry]
name = "lesson4-step2"
version = "0.1.0"
description = "FastAPI service for ONNX model inference"
authors = ["MLOps Course"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pillow = "^10.0.0"
onnxruntime = "^1.16.0"
transformers = "^4.35.0"
torch = "^2.1.0"
numpy = "^1.24.0"
pydantic = "^2.5.0"
requests = "^2.31.0"
PyYAML = "^6.0"
colorama = "^0.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
httpx = "^0.25.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
